+++ apps/management/commands/create_dummy_admin.py
@@ -6,11 +6,11 @@ from django.db.utils import IntegrityError
 
 
 class Command(BaseCommand):
-    help = ('Create a TECHNICAL Administrator account for testing and debugging purposes. '
+    help = ('Create a TECHNICAL administrator account for testing and debugging purposes. '
             'The command is NOT intended to run in Production and will not work there.')
 
     def handle(self, *args: str, **options: str) -> None:
-        """Command handler method. Creates a TECHNICAL superuser with login 'admin' and password 'admin'.
+        """Create a TECHNICAL superuser with login 'admin' and password 'admin'.
 
         Args:
             args: positional command arguments (not used, interface requirement).
@@ -19,9 +19,11 @@ class Command(BaseCommand):
         Raises:
             CommandError if the command is run in a production environment.
         """
-        if os.environ['DJANGO_SETTINGS_MODULE'] == 'stubborn.settings.production':
-            raise CommandError('This command cannot be run in a production environment! '
-                               'Use the "createsuperuser" command to create a superuser in a production environment.')
+        if os.environ['DJANGO_SETTINGS_MODULE'] != 'stubborn.settings.local':
+            raise CommandError(
+                'This command can run in a local environment only! Use the "create_admin" or "createsuperuser" '
+                'command to create a superuser in a production or staging environment.'
+            )
         username = 'admin'
         email = 'admin@example.com'
         password = 'admin'